feat: Remove ApiService and update base URL
This commit removes the `ApiService` class, which was responsible for making network requests to a placeholder API. The base URL in `ApiConstants` has been updated to reflect the actual backend service. Additionally, the `MyDdocDocScreen` has been renamed to `DoctorFutureScreen`. The API-related logic in `HomeViewModel` and the usage of `apiResult` in `MyPageScreen` and `HomeScreen` have also been removed.